Fix any

+++ src/org/pixelgame/Engine/object/Mob.java
@@ -2,7 +2,7 @@ package org.pixelgame.Engine.object;
 
 public class Mob extends Sprite {
     protected float runSpeed = 50.0f;
-    public Mob(int id, float posX, float posY) {
+    public Mob(int id, int posX, int posY) {
         super(id, posX, posY);
     }
 }
